Added controller concerns for devise params
