+++ app/controllers/application_controller.rb
@@ -1,10 +1,4 @@
 class ApplicationController < ActionController::Base
     protect_from_forgery with: :exception
-    
-    before_action :configure_permitted_parameters, if: :devise_controller?
-
-    def configure_permitted_parameters
-        devise_parameter_sanitizer.permit(:sign_up, keys: [:name])
-        devise_parameter_sanitizer.permit(:account_update, keys: [:name])
-    end
-end
+    include DeviseWhitelist
+  end
