--- app/controllers/concerns/placeholder.rb
@@ -1,7 +1,0 @@
-module Placeholder
-    extend ActiveSupport::Concern
-  
-    def self.image_generator(height:, width:)
-      "http://placehold.it/#{height}x#{width}"
-    end
-  end
